% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSR_citations.R
\name{NSR_citations}
\alias{NSR_citations}
\title{Get citation information}
\usage{
NSR_citations()
}
\value{
Dataframe containing bibtex-formatted citation information
}
\description{
Returns information needed to cite the NSR
}
\note{
This function provides citation information in bibtex format that can be used with reference manager software (e.g. Paperpile, Zotero). Please do remember to cite both the sources and the NSR, as the NSR couldn't exist without these sources!
}
\examples{
{
citation_info <- NSR_citations()
}

}
