% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSR_from_coordinates.R
\name{NSR_from_coordinates}
\alias{NSR_from_coordinates}
\title{Check the native status for an occurrence using coordinates}
\usage{
NSR_from_coordinates(occurrence_dataframe, ...)
}
\arguments{
\item{occurrence_dataframe}{A properly formatted dataframe. Should contain the columns: species, latitude, and longitude, and ID in that order.}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
Dataframe containing NSR results.
}
\description{
NSR returns information on native status for species based on their latitude and longitude.
}
\examples{
\dontrun{
library(BIEN)

# Get some data

ar_occs <- BIEN_occurrence_species(species = "Acer rubrum",
                                  natives.only = FALSE,
                                  political.boundaries = TRUE,
                                  limit=100)


# Rename columns so that we can differentiate between coordinates declared
# and coordinates inferred

colnames(ar_occs)[which(colnames(ar_occs)=="country")] <- "country_declared"
colnames(ar_occs)[which(colnames(ar_occs)=="state_province")] <- "state_declared"
colnames(ar_occs)[which(colnames(ar_occs)=="county")] <- "county_declared"

# Select needed fields and add an ID column

ar_occs_subset <- ar_occs[c("scrubbed_species_binomial","latitude","longitude")]

# Toss duplicates (speeds up processing)

ar_occs_subset <- unique(ar_occs_subset)

# Add an id column

ar_occs_subset$id <- as.character(1:nrow(ar_occs_subset))

# Get results

results_ar <- NSR_from_coordinates(occurrence_dataframe = ar_occs_subset)

# Join results back to original data

results_ar <- ar_occs_subset |>
 merge(y = results_ar,by.x = c("id"),
       by.y = c("input_id"))

ar_occs <- ar_occs |> merge(results_ar,
                           by = c("latitude","longitude","scrubbed_species_binomial"))
}
}
