% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_cache.R
\name{NSR_local_remove}
\alias{NSR_local_remove}
\title{Remove the local NSR tables}
\usage{
NSR_local_remove(dir = nsr_cache_dir(), quiet = FALSE)
}
\arguments{
\item{dir}{Cache directory.}

\item{quiet}{Suppress messages?}
}
\value{
Invisibly, the files removed.
}
\description{
Remove the local NSR tables
}
