% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSR_political_divisions.R
\name{NSR_political_divisions}
\alias{NSR_political_divisions}
\title{Get information on political divisions with checklists within the NSR}
\usage{
NSR_political_divisions(by_country = TRUE, ...)
}
\arguments{
\item{...}{Additional parameters passed to internal functions.}

\item{country}{(optional) Character. Limits results to political divisions containing checklist information for a single country.}

\item{checklist}{If TRUE (the default) limits the result to political divisions represented by one or more comprehensive checklists.}
}
\value{
data.frame containing information on political divisions within the NSR database.
}
\description{
NSR_political_divisions returns information on political divisions with checklist information present in the NSR.
}
\note{
Setting checklist to FALSE returns a list of political divisions that can be used to standardize spellings.
}
\examples{
\dontrun{

#To get a list of all political divisions with comprehensive checklists:
checklists_per_country <- NSR_political_divisions()

#To get a list of all checklists the associated countries, set "by_country" to FALSE
countries_per_checklist <- NSR_political_divisions(by_country=FALSE)

}
}
