% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSR_data_dictionary.R
\name{NSR_data_dictionary}
\alias{NSR_data_dictionary}
\title{Get NSR data dictionary}
\usage{
NSR_data_dictionary(native_status = FALSE, ...)
}
\arguments{
\item{native_status}{Logical. If FALSE(Default) returns information on fields. If TRUE, returns information on Native Status categories.}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
Data.frame containing requested data dictionary contents.
}
\description{
Returns information from the NSR data dictionary
}
\examples{
{
NSR_fields <- NSR_data_dictionary()

status_codes <- NSR_data_dictionary(native_status = TRUE)
}

}
