% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSR_sources.R
\name{NSR_sources}
\alias{NSR_sources}
\title{Get information on sources used by the NSR}
\usage{
NSR_sources(...)
}
\value{
Dataframe containing information about the sources used in the current NSR version.
}
\description{
Return metadata about the current NSR sources
}
\examples{
{
sources <- NSR_sources()
}

}
